requirements.txt
requests>=2.28
Pillow>=10.0
beautifulsoup4>=4.12
PyYAML>=6.0
schedule>=1.2
pytest>=7.0
rumps>=0.4.0; sys_platform == "darwin"
